+++ energyquantified/utils/__init__.py
@@ -1,10 +1,7 @@
 from .page import Page
-from .lists import PeriodseriesList, TimeseriesList
 from .misc import dict_to_str
 
 __all__ = [
     "Page",
-    "PeriodseriesList",
-    "TimeseriesList",
     "dict_to_str",
 ]